Trying to use recycler list
no success yet
